+++ engrama/logger.py
@@ -31,7 +31,7 @@ def get_logger(name: str) -> logging.Logger:
             level = os.getenv("ENGRAMA_LOG_LEVEL", "INFO").upper()
             logger.setLevel(getattr(logging, level, logging.INFO))
 
-            handler = logging.StreamHandler(sys.stdout)
+            handler = logging.StreamHandler(sys.stderr)
             handler.setLevel(getattr(logging, level, logging.INFO))
 
             formatter = logging.Formatter(